+++ apps/explorer/app/contract/[id]/opengraph-image.tsx
@@ -6,7 +6,7 @@ import { humanBytes, humanDate } from '@siafoundation/units'
 import { getOGImage } from '../../../components/OGImageEntity'
 import { siaCentralApi } from '../../../config'
 import { truncate } from '@siafoundation/design-system'
-import { lowerCase } from 'lodash'
+import { lowerCase } from 'lodash-es'
 import { siacoinToFiat } from '../../../lib/currency'
 import { CurrencyOption, currencyOptions } from '@siafoundation/react-core'