+++ libs/design-system/src/app/TransactionDetailsDialog.tsx
@@ -7,7 +7,7 @@ import { Dialog } from '../core/Dialog'
 import { getTitleId } from '../lib/utils'
 import { Transaction } from '@siafoundation/types'
 import { getTxTypeLabel, TxType } from '../lib/entityTypes'
-import { upperFirst } from 'lodash'
+import { upperFirst } from 'lodash-es'
 
 type Props = {
   id: string