+++ libs/react-core/src/useAppSettings/useGpuFeatures.tsx
@@ -1,7 +1,9 @@
 import { useCallback, useEffect, useState } from 'react'
 import useLocalStorageState from 'use-local-storage-state'
-import { getGPUTier } from 'detect-gpu'
+import * as detectGpu from 'detect-gpu'
 import { usePrefersReducedMotion } from '../userPrefersReducedMotion'
+// esm compat
+const { getGPUTier } = detectGpu
 
 export function useGpuFeatures() {
   const reduceMotion = usePrefersReducedMotion()