+++ libs/fonts/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @siafoundation/fonts
 
+## 0.1.5
+
+### Patch Changes
+
+- esm lodash support.
+
+## 0.1.4
+
+### Patch Changes
+
+- lodash esm.
+
+## 0.1.3
+
+### Patch Changes
+
+- lodash es.
+
+## 0.1.2
+
+### Patch Changes
+
+- 4294302f: Adjust packaging.
+
 ## 0.1.0
 
 ### Minor Changes